Multi machine (#93)
Allow multiple VMs to be assigned to a single clickpoint. 
Users can do a wildcard match: win10.user.* 
or supply a range:  ubuntu.user.[5,10]
The wildcard/range can currently only appear at the end of a VM name, but that could change. If we're going to allow them in arbitrary locations, though, I think we might as well just allow the user to enter regular expressions. 
Internal ticket: CRU-1147